auth context

+++ frontend/Mern-Project/src/components/WorkoutDetails.jsx
@@ -1,12 +1,20 @@
+import { useAuthContext } from "../hooks/useAuthContext";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 export const WorkoutDetails = ({ workout })=> {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext(); // This hook will ensure that the UserContext is available throughout the component tree.
     const handleClick = async () => {
+        if(!user){
+            return
+        }
         const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
             method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         if(!response.ok) {
             console.error('Failed to delete workout');
